Refactoring to a general utils file
