+++ dev/breeze/src/airflow_breeze/params/doc_build_params.py
@@ -20,21 +20,7 @@ import os
 from dataclasses import dataclass
 
 from airflow_breeze.branch_defaults import AIRFLOW_BRANCH
-
-providers_prefix = "apache-airflow-providers-"
-
-
-def get_provider_name_from_short_hand(short_form_providers: tuple[str]):
-    providers = []
-    for short_form_provider in short_form_providers:
-        if short_form_provider == "providers-index":
-            providers.append("apache-airflow-providers")
-            continue
-
-        short_form_provider.split(".")
-        parts = "-".join(short_form_provider.split("."))
-        providers.append(providers_prefix + parts)
-    return tuple(providers)
+from airflow_breeze.utils.general_utils import get_provider_name_from_short_hand
 
 
 @dataclass
