+++ dev/breeze/src/airflow_breeze/utils/publish_docs_helpers.py
@@ -26,26 +26,12 @@ from typing import Any
 
 import yaml
 
+from airflow_breeze.utils.general_utils import get_provider_name_from_short_hand
+
 CONSOLE_WIDTH = 180
 
 ROOT_DIR = Path(__file__).parents[5].resolve()
 PROVIDER_DATA_SCHEMA_PATH = ROOT_DIR / "airflow" / "provider.yaml.schema.json"
-
-providers_prefix = "apache-airflow-providers-"
-
-
-def get_provider_name_from_short_hand(short_form_providers: tuple[str]):
-    providers = []
-    for short_form_provider in short_form_providers:
-        short_form_provider.split(".")
-        if short_form_provider == "providers-index":
-            providers.append("apache-airflow-providers")
-            continue
-
-        short_form_provider.split(".")
-        parts = "-".join(short_form_provider.split("."))
-        providers.append(providers_prefix + parts)
-    return tuple(providers)
 
 
 def _load_schema() -> dict[str, Any]:
